[4.3] Decompose session approval into private method
Correct my interpretation of the requirements in that only one of the sessions must be approved whereas my first understanding was that all sessions must be approved. Break this approval process into a separate function and make the conditional check on the mutated collection to separate the act of approval with the checking of business logic rules of registering a speaker.